+++ src/Controller/Admin/LocationController.php
@@ -6,6 +6,7 @@ use App\Entity\Location;
 use App\Form\LocationType;
 use App\Repository\LocationRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -68,7 +69,7 @@ class LocationController extends AbstractController
      *
      * @Route("/edit/{id}", name="admin_location_edit")
      */
-    public function edit(Request $request, TranslatorInterface $translator, Location $location): Response
+    public function edit(Request $request, Location $location): Response
     {
         $form = $this->createForm(LocationType::class, $location);
         $form->handleRequest($request);
@@ -78,16 +79,13 @@ class LocationController extends AbstractController
             $em->persist($location);
             $em->flush();
 
-            $translated = $translator->trans('flash.success.edit');
-            $this->addFlash(
-                'success',
-                $translated.'.'
-            );
+            $this->addFlash('success', 'flash.success.edit');
         }
 
         return $this->render('admin/location/edit.html.twig', [
             'location' => $location,
             'form' => $form->createView(),
+            'deleteForm' => $this->createDeleteForm($location)->createView(),
         ]);
     }
 
@@ -96,20 +94,31 @@ class LocationController extends AbstractController
      *
      * @Route("/delete/{id}", name="admin_location_delete")
      */
-    public function delete(TranslatorInterface $translator, Location $location): RedirectResponse
+    public function delete(Request $request, TranslatorInterface $translator, Location $location): RedirectResponse
     {
-        $em = $this->getDoctrine()->getManager();
+        $form = $this->createDeleteForm($location);
+        $form->handleRequest($request);
 
-        $translated = str_replace('%location%', $location->getName(), $translator->trans('flash.success.deleted.location'));
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
 
-        $em->remove($location);
-        $em->flush();
+            $em->remove($location);
+            $em->flush();
 
-        $this->addFlash(
-            'success',
-            $translated.'.'
-        );
+            $translated = str_replace('%location%', $location->getName(), $translator->trans('flash.success.deleted.location'));
+            $this->addFlash('success', $translated);
+        } else {
+            $this->addFlash('danger', 'flash.error.deleted');
+        }
 
         return $this->redirectToRoute('admin_index');
+    }
+
+    private function createDeleteForm(Location $location): FormInterface
+    {
+        return $this->createFormBuilder()
+            ->setAction($this->generateUrl('admin_location_delete', ['id' => $location->getId()]))
+            ->setMethod('DELETE')
+            ->getForm();
     }
 }
